+++ utils/filterProviderKeys.js
@@ -1,0 +1,16 @@
+const filterProviderKeys = (keys, filter) => {
+  if (!filter) return keys;
+
+  const filterFunc = (key, filter, exclude) => {
+    if (!filter) return true;
+
+    const filterIncludes = filter.includes(key);
+    return exclude ? !filterIncludes : filterIncludes;
+  };
+
+  return keys
+    .filter(key => filterFunc(key, filter.include))
+    .filter(key => filterFunc(key, filter.exclude, true));
+};
+
+module.exports = filterProviderKeys
